requirements.txt
django == 3.2.1
requests == 2.25.1
pillow == 8.2.0
djangorestframework == 3.12.4